Support Table instance where table string is supported
The Table class supports aliases and schemas, but the query builder
and syntax builder required string parameters instead.

The easiest way to add support for table alias and multiple database
schema queries without changing the API, is to allow Table objects
to be passed to setTable() and related methods.

Also:

- docs: Fix `@return` of AbstractBaseQuery::getTable.

- docs: Fix `@param $table` of AbstractBaseQuery::orderBy.

- minor: Remove redundant `array()` wrap on table names
  from AbstractBaseQuery to SyntaxFactory::createTable, and
  from Column to SyntaxFactory::createTable.

- Column: Fix bug where `(string) $table` caused alias and schema
  to be lost, due to unexpected Table being casted to plain name
  via Table::__toString.

Ref https://github.com/nilportugues/php-sql-query-builder/issues/93.